Move all flag formatters to the header file
